requirements.txt
fastapi
uvicorn
firebase-admin
sentence-transformers
pinecone-client
langchain-google-genai
google-generativeai
pydantic
requests
python-dotenv
pinecone